backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
snowflake-connector-python==3.13.1
google-genai==1.73.1
elevenlabs==1.3.0
httpx==0.27.0
pydantic==2.7.1
